+++ android/src/main/java/com/cyanchill/missingcore/metadataretriever/Exceptions.kt
@@ -1,4 +1,4 @@
-package com.missingcore.metadataretriever
+package com.cyanchill.missingcore.metadataretriever
 
 
 class TrackGroupArrayException: