+++ android/src/main/java/com/cyanchill/missingcore/metadataretriever/MediaMetadataRetrieverUtils.kt
@@ -1,4 +1,4 @@
-package com.missingcore.metadataretriever
+package com.cyanchill.missingcore.metadataretriever
 
 import android.media.MediaMetadataRetriever
 import com.facebook.react.bridge.ReactApplicationContext